Update the website's structure and add a blog section
This commit includes structural updates to the existing website, moving stylesheets and scripts to respective `styles` and `scripts` directories. The old link references have also been updated to point to the correct locations. Moreover, the Blog section has been added to the website, containing placeholder content for the time being. It also reorganizes navbar links and minor adjustments in styles.